feat: 돋보기를 눌렀을 때 프로그램 정보 Dialog가 나올 수 있도록 수정
